requirements.txt
altair==5.0.1
altgraph==0.17.2
appscript==1.2.0
asgiref==3.4.1
attrs==23.1.0
auto-py-to-exe==2.22.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.1
blinker==1.6.2
bottle==0.12.23
bottle-websocket==0.2.9
brython==3.10.6
bs4==0.0.1
cachetools==5.3.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.3
Django==3.2.8
dnspython==2.3.0
docutils==0.18.1
docx==0.2.4
docx2pdf==0.1.8
docxcompose==1.3.4
docxtpl==0.16.0
EasyProcess==1.1
Eel==0.14.0
et-xmlfile==1.1.0
Faker==19.2.0
favicon==0.7.0
filelock==3.3.1
Flask==2.3.2
fonttools==4.41.1
fpdf==1.7.2
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.32
greenlet==1.1.3
htbuilder==0.6.1
idna==3.3
image==1.5.33
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
lxml==4.7.1
macholib==1.16
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.1
matplotlib==3.7.2
mdurl==0.1.2
modulegraph==0.19.2
more-itertools==10.0.0
MouseInfo==0.1.3
numpy==1.21.3
opencv-python==4.6.0.66
openpyxl==3.0.9
packaging==23.1
pandas==1.3.4
Pillow==9.5.0
pipenv==2021.5.29
platformdirs==2.4.0
protobuf==4.23.4
py2app==0.28
pyarrow==12.0.1
PyAutoGUI==0.9.53
pydeck==0.8.0
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
pymdown-extensions==10.1
pymongo==4.3.3
Pympler==1.0.1
PyMsgBox==1.0.9
PyMuPDF==1.21.1
pyobjc==7.3
pyobjc-core==7.3
pyobjc-framework-Accessibility==7.3
pyobjc-framework-Accounts==7.3
pyobjc-framework-AddressBook==7.3
pyobjc-framework-AdServices==7.3
pyobjc-framework-AdSupport==7.3
pyobjc-framework-AppleScriptKit==7.3
pyobjc-framework-AppleScriptObjC==7.3
pyobjc-framework-ApplicationServices==7.3
pyobjc-framework-AppTrackingTransparency==7.3
pyobjc-framework-AuthenticationServices==7.3
pyobjc-framework-AutomaticAssessmentConfiguration==7.3
pyobjc-framework-Automator==7.3
pyobjc-framework-AVFoundation==7.3
pyobjc-framework-AVKit==7.3
pyobjc-framework-BusinessChat==7.3
pyobjc-framework-CalendarStore==7.3
pyobjc-framework-CallKit==7.3
pyobjc-framework-CFNetwork==7.3
pyobjc-framework-ClassKit==7.3
pyobjc-framework-CloudKit==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Collaboration==7.3
pyobjc-framework-ColorSync==7.3
pyobjc-framework-Contacts==7.3
pyobjc-framework-ContactsUI==7.3
pyobjc-framework-CoreAudio==7.3
pyobjc-framework-CoreAudioKit==7.3
pyobjc-framework-CoreBluetooth==7.3
pyobjc-framework-CoreData==7.3
pyobjc-framework-CoreHaptics==7.3
pyobjc-framework-CoreLocation==7.3
pyobjc-framework-CoreMedia==7.3
pyobjc-framework-CoreMediaIO==7.3
pyobjc-framework-CoreMIDI==7.3
pyobjc-framework-CoreML==7.3
pyobjc-framework-CoreMotion==7.3
pyobjc-framework-CoreServices==7.3
pyobjc-framework-CoreSpotlight==7.3
pyobjc-framework-CoreText==7.3
pyobjc-framework-CoreWLAN==7.3
pyobjc-framework-CryptoTokenKit==7.3
pyobjc-framework-DeviceCheck==7.3
pyobjc-framework-DictionaryServices==7.3
pyobjc-framework-DiscRecording==7.3
pyobjc-framework-DiscRecordingUI==7.3
pyobjc-framework-DiskArbitration==7.3
pyobjc-framework-DVDPlayback==7.3
pyobjc-framework-EventKit==7.3
pyobjc-framework-ExceptionHandling==7.3
pyobjc-framework-ExecutionPolicy==7.3
pyobjc-framework-ExternalAccessory==7.3
pyobjc-framework-FileProvider==7.3
pyobjc-framework-FileProviderUI==7.3
pyobjc-framework-FinderSync==7.3
pyobjc-framework-FSEvents==7.3
pyobjc-framework-GameCenter==7.3
pyobjc-framework-GameController==7.3
pyobjc-framework-GameKit==7.3
pyobjc-framework-GameplayKit==7.3
pyobjc-framework-ImageCaptureCore==7.3
pyobjc-framework-IMServicePlugIn==7.3
pyobjc-framework-InputMethodKit==7.3
pyobjc-framework-InstallerPlugins==7.3
pyobjc-framework-InstantMessage==7.3
pyobjc-framework-Intents==7.3
pyobjc-framework-IOSurface==7.3
pyobjc-framework-iTunesLibrary==7.3
pyobjc-framework-KernelManagement==7.3
pyobjc-framework-LatentSemanticMapping==7.3
pyobjc-framework-LaunchServices==7.3
pyobjc-framework-libdispatch==7.3
pyobjc-framework-LinkPresentation==7.3
pyobjc-framework-LocalAuthentication==7.3
pyobjc-framework-MapKit==7.3
pyobjc-framework-MediaAccessibility==7.3
pyobjc-framework-MediaLibrary==7.3
pyobjc-framework-MediaPlayer==7.3
pyobjc-framework-MediaToolbox==7.3
pyobjc-framework-Metal==7.3
pyobjc-framework-MetalKit==7.3
pyobjc-framework-MetalPerformanceShaders==7.3
pyobjc-framework-MetalPerformanceShadersGraph==7.3
pyobjc-framework-MLCompute==7.3
pyobjc-framework-ModelIO==7.3
pyobjc-framework-MultipeerConnectivity==7.3
pyobjc-framework-NaturalLanguage==7.3
pyobjc-framework-NetFS==7.3
pyobjc-framework-Network==7.3
pyobjc-framework-NetworkExtension==7.3
pyobjc-framework-NotificationCenter==7.3
pyobjc-framework-OpenDirectory==7.3
pyobjc-framework-OSAKit==7.3
pyobjc-framework-OSLog==7.3
pyobjc-framework-PassKit==7.3
pyobjc-framework-PencilKit==7.3
pyobjc-framework-Photos==7.3
pyobjc-framework-PhotosUI==7.3
pyobjc-framework-PreferencePanes==7.3
pyobjc-framework-PushKit==7.3
pyobjc-framework-Quartz==7.3
pyobjc-framework-QuickLookThumbnailing==7.3
pyobjc-framework-ReplayKit==7.3
pyobjc-framework-SafariServices==7.3
pyobjc-framework-SceneKit==7.3
pyobjc-framework-ScreenSaver==7.3
pyobjc-framework-ScreenTime==7.3
pyobjc-framework-ScriptingBridge==7.3
pyobjc-framework-SearchKit==7.3
pyobjc-framework-Security==7.3
pyobjc-framework-SecurityFoundation==7.3
pyobjc-framework-SecurityInterface==7.3
pyobjc-framework-ServiceManagement==7.3
pyobjc-framework-Social==7.3
pyobjc-framework-SoundAnalysis==7.3
pyobjc-framework-Speech==7.3
pyobjc-framework-SpriteKit==7.3
pyobjc-framework-StoreKit==7.3
pyobjc-framework-SyncServices==7.3
pyobjc-framework-SystemConfiguration==7.3
pyobjc-framework-SystemExtensions==7.3
pyobjc-framework-UniformTypeIdentifiers==7.3
pyobjc-framework-UserNotifications==7.3
pyobjc-framework-UserNotificationsUI==7.3
pyobjc-framework-VideoSubscriberAccount==7.3
pyobjc-framework-VideoToolbox==7.3
pyobjc-framework-Virtualization==7.3
pyobjc-framework-Vision==7.3
pyobjc-framework-WebKit==7.3
pyparsing==3.0.9
PyPDF2==2.10.3
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
PySimpleGUI==4.53.0
python-dateutil==2.8.2
python-docx==0.8.11
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==0.2.4
PyYAML==6.0.1
referencing==0.30.0
reportlab==4.0.4
requests==2.28.0
rich==13.5.0
rpds-py==0.9.2
rubicon-objc==0.4.1
selenium==3.141.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.15
sqlparse==0.4.2
st-annotated-text==4.0.0
streamlit==1.25.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.64.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.9
validators==0.20.0
virtualenv==20.15.1
virtualenv-clone==0.5.7
Werkzeug==2.3.4
whichcraft==0.6.1
zipp==3.16.2
zope.event==4.5.0
zope.interface==5.4.0
